Updated plugins and delete unneccessary plugins

+++ plugins/backwpup/inc/class-destination-folder.php
@@ -11,8 +11,11 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 	public function option_defaults() {
 
 		$upload_dir = wp_upload_dir();
+		$backups_dir = trailingslashit( str_replace( '\\', '/',$upload_dir[ 'basedir' ] ) ) . 'backwpup-' . BackWPup::get_plugin_data( 'hash' ) . '-backups/';
+		$content_path = trailingslashit( str_replace( '\\', '/', WP_CONTENT_DIR ) );
+		$backups_dir = str_replace( $content_path, '', $backups_dir );
 
-		return array( 'maxbackups' => 15, 'backupdir' => trailingslashit( str_replace( '\\', '/',$upload_dir[ 'basedir' ] ) ) . 'backwpup-' . BackWPup::get_plugin_data( 'hash' ) . '-backups/', 'backupsyncnodelete' => TRUE );
+		return array( 'maxbackups' => 15, 'backupdir' => $backups_dir, 'backupsyncnodelete' => TRUE );
 	}
 
 
@@ -57,11 +60,7 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 	 */
 	public function edit_form_post_save( $jobid ) {
 
-		$_POST[ 'backupdir' ] = trailingslashit( str_replace( '//', '/', str_replace( '\\', '/', trim( stripslashes( $_POST[ 'backupdir' ] ) ) ) ) );
-		if ( $_POST[ 'backupdir' ][ 0 ] == '.' || ( $_POST[ 'backupdir' ][ 0 ] != '/' && ! preg_match( '#^[a-zA-Z]:/#', $_POST[ 'backupdir' ] ) ) )
-			$_POST[ 'backupdir' ] = trailingslashit( str_replace( '\\', '/', ABSPATH ) ) . $_POST[ 'backupdir' ];
-		if ( $_POST[ 'backupdir' ] == '/' )
-			$_POST[ 'backupdir' ] = '';
+		$_POST[ 'backupdir' ] = trailingslashit( str_replace( array( '//', '\\' ), '/', trim( stripslashes( $_POST[ 'backupdir' ] ) ) ) );
 		BackWPup_Option::update( $jobid, 'backupdir', $_POST[ 'backupdir' ] );
 
 		BackWPup_Option::update( $jobid, 'maxbackups', isset( $_POST[ 'maxbackups' ] ) ? (int)$_POST[ 'maxbackups' ] : 0 );
@@ -74,9 +73,20 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 	 */
 	public function file_delete( $jobdest, $backupfile ) {
 
-		if ( is_writeable( $backupfile ) && !is_dir( $backupfile ) && !is_link( $backupfile ) )
-			 unlink( $backupfile );
+		list( $jobid, $dest ) = explode( '_', $jobdest, 2 );
 
+		if ( empty( $jobid ) ) {
+			return;
+		}
+
+		$backup_dir = esc_attr( BackWPup_Option::get( (int)$jobid, 'backupdir' ) );
+		$backup_dir = BackWPup_File::get_absolute_path( $backup_dir );
+
+		$backupfile =  realpath( trailingslashit( $backup_dir ) . basename( $backupfile ) );
+
+		if ( $backupfile && is_writeable( $backupfile ) && !is_dir( $backupfile ) && !is_link( $backupfile ) ) {
+			 unlink( $backupfile );
+		}
 	}
 
 	/**
@@ -85,7 +95,13 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 	 */
 	public function file_download( $jobid, $get_file ) {
 
-		if ( is_readable( $get_file ) ) {
+		$backup_dir = esc_attr( BackWPup_Option::get( (int)$jobid, 'backupdir' ) );
+		$backup_dir = BackWPup_File::get_absolute_path( $backup_dir );
+
+		$get_file = realpath( trailingslashit( $backup_dir ) . basename( $get_file ) );
+
+		if ( $get_file && is_readable( $get_file ) ) {
+			while( @ob_end_clean() );
 			header( "Pragma: public" );
 			header( "Expires: 0" );
 			header( "Cache-Control: must-revalidate, post-check=0, pre-check=0" );
@@ -94,14 +110,12 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 			header( "Content-Transfer-Encoding: binary" );
 			header( "Content-Length: " . filesize( $get_file ) );
 			@set_time_limit( 300 );
-			//chunked readfile
-			@ob_end_clean();
+			//chunked read file
 			$handle = fopen( $get_file, 'rb' );
 			if ( $handle ) {
 				while ( ! feof( $handle ) ) {
-					echo fread( $handle, 20482048 ); //2MB chunkes
-					@ob_flush();
-					@flush();
+					echo fread( $handle, 10241024 ); //chunks
+					flush();
 				}
 				fclose( $handle );
 			}
@@ -120,10 +134,11 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 	 */
 	public function file_get_list( $jobdest ) {
 
-		list( $jobid, $dest ) = explode( '_', $jobdest );
+		list( $jobid, $dest ) = explode( '_', $jobdest, 2 );
 		$filecounter    = 0;
 		$files          = array();
 		$backup_folder  = BackWPup_Option::get( $jobid, 'backupdir' );
+		$backup_folder  = BackWPup_File::get_absolute_path( $backup_folder );
 		if ( is_dir( $backup_folder ) && $dir = opendir( $backup_folder ) ) { //make file list
 			while ( FALSE !== ( $file = readdir( $dir ) ) ) {
 				if ( in_array( $file, array( '.', '..', 'index.php', '.htaccess', '.donotbackup' ) ) || is_dir( $backup_folder . $file ) || is_link( $backup_folder . $file ) )
@@ -136,7 +151,7 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 					$files[ $filecounter ][ 'downloadurl' ] = add_query_arg( array(
 																				  'page'   => 'backwpupbackups',
 																				  'action' => 'downloadfolder',
-																				  'file'   => $backup_folder . $file,
+																				  'file'   => $file,
 																				  'jobid'  => $jobid
 																			 ), network_admin_url( 'admin.php' ) );
 					$files[ $filecounter ][ 'filesize' ]    = filesize( $backup_folder . $file );
@@ -144,7 +159,7 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 					$filecounter ++;
 				}
 			}
-			@closedir( $dir );
+			closedir( $dir );
 		}
 
 		return $files;
@@ -154,19 +169,20 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 	 * @param $job_object
 	 * @return bool
 	 */
-	public function job_run_archive( &$job_object ) {
+	public function job_run_archive( BackWPup_Job$job_object ) {
 
 		$job_object->substeps_todo = 1;
 		if ( ! empty( $job_object->job[ 'jobid' ] ) )
 			BackWPup_Option::update( $job_object->job[ 'jobid' ], 'lastbackupdownloadurl', add_query_arg( array(
 																								  'page'   => 'backwpupbackups',
 																								  'action' => 'downloadfolder',
-																								  'file'   => $job_object->backup_folder . $job_object->backup_file
+																								  'file'   => basename( $job_object->backup_file ),
+																								  'jobid'  => $job_object->job[ 'jobid' ]
 																							 ), network_admin_url( 'admin.php' ) ) );
 		//Delete old Backupfiles
 		$backupfilelist = array();
 		$files          = array();
-		if ( $dir = opendir( $job_object->backup_folder ) ) { //make file list
+		if ( is_writable( $job_object->backup_folder ) && $dir = opendir( $job_object->backup_folder ) ) { //make file list
 			while ( FALSE !== ( $file = readdir( $dir ) ) ) {
 				if ( is_writeable( $job_object->backup_folder . $file ) && ! is_dir( $job_object->backup_folder . $file ) && ! is_link( $job_object->backup_folder . $file ) ) {
 					//list for deletion
@@ -174,7 +190,7 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 						$backupfilelist[ filemtime( $job_object->backup_folder . $file ) ] = $file;
 				}
 			}
-			@closedir( $dir );
+			closedir( $dir );
 		}
 		if ( $job_object->job[ 'maxbackups' ] > 0 ) {
 			if ( count( $backupfilelist ) > $job_object->job[ 'maxbackups' ] ) {
@@ -185,8 +201,9 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 						break;
 					unlink( $job_object->backup_folder . $file );
 					foreach ( $files as $key => $filedata ) {
-						if ( $filedata[ 'file' ] == $job_object->backup_folder . $file )
+						if ( $filedata[ 'file' ] == $job_object->backup_folder . $file ) {
 							unset( $files[ $key ] );
+						}
 					}
 					$numdeltefiles ++;
 				}
@@ -201,12 +218,12 @@ class BackWPup_Destination_Folder extends BackWPup_Destinations {
 	}
 
 	/**
-	 * @param $job_object
+	 * @param $job_settings array
 	 * @return bool
 	 */
-	public function can_run( $job_object ) {
+	public function can_run( array $job_settings ) {
 
-		if ( empty( $job_object->job[ 'backupdir' ] ) || $job_object->job[ 'backupdir' ] == '/' )
+		if ( empty( $job_settings[ 'backupdir' ] ) || $job_settings[ 'backupdir' ] == '/' )
 			return FALSE;
 
 		return TRUE;
